vnda-docs-mcp-client/pyproject.toml
[project]
name = "vnda-docs-mcp-client"
version = "0.1.0"
description = "Cliente/orquestrador: recebe perguntas de bots (ex.: Sebastião), usa MCP (doc Vnda) + OpenAI ChatGPT e devolve resposta. Reutilizável por outros bots; cada bot habilita via ORCHESTRATOR_URL."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "openai>=1.12",
    "mcp[cli]>=1.0",
    "httpx>=0.27",
    "httpx-sse>=0.4",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = []

[tool.hatch.build.targets.wheel]
packages = ["orchestrator"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
